edit pre-emphasis value in text box (#20)
Previously, only the slider worked to edit the pre-emphasis value. This change allows you to edit the value by typing it into the text box, then pressing "ENTER" on your keyboard. The slider and textbox should now be synced, so that the desired pre-emphasis value is maintained if you switch between using the slider or text box to edit the value. Additionally, the slider position should reflect changes in value regardless of how the value is changed. 

Next, see if there are any potential performance issues.